+++ src/pid.cpp
@@ -1,6 +1,6 @@
 #include "pid.hpp"
 
-float PIDController::calculate(float set_point, float current_pos) {
+double PIDController::calculate(double set_point, double current_pos) {
 	error = current_pos - set_point;
 	error_change = error - error_prev;
 	error_total += error;
@@ -9,13 +9,13 @@ float PIDController::calculate(float set_point, float current_pos) {
 	return (error * kP + error_total * kI + error_change * kD);
 }
 
-void PIDController::set_gains(float kP, float kI, float kD) {
+void PIDController::set_gains(double kP, double kI, double kD) {
 	this->kP = kP;
 	this->kI = kI;
 	this->kD = kD;
 }
 
-float PIDController::get_error() { return error; }
+double PIDController::get_error() { return error; }
 
 void PIDController::reset() {
 	error_change = 0;
